+++ internal/dnsforward/msg.go
@@ -50,7 +50,7 @@ func (s *Server) genDNSFilterMessage(
 	req := dctx.Req
 	qt := req.Question[0].Qtype
 	if qt != dns.TypeA && qt != dns.TypeAAAA {
-		if s.conf.BlockingMode == BlockingModeNullIP {
+		if s.dnsFilter.BlockingMode == filtering.BlockingModeNullIP {
 			return s.makeResponse(req)
 		}
 
@@ -59,9 +59,9 @@ func (s *Server) genDNSFilterMessage(
 
 	switch res.Reason {
 	case filtering.FilteredSafeBrowsing:
-		return s.genBlockedHost(req, s.conf.SafeBrowsingBlockHost, dctx)
+		return s.genBlockedHost(req, s.dnsFilter.SafeBrowsingBlockHost, dctx)
 	case filtering.FilteredParental:
-		return s.genBlockedHost(req, s.conf.ParentalBlockHost, dctx)
+		return s.genBlockedHost(req, s.dnsFilter.ParentalBlockHost, dctx)
 	case filtering.FilteredSafeSearch:
 		// If Safe Search generated the necessary IP addresses, use them.
 		// Otherwise, if there were no errors, there are no addresses for the
@@ -76,13 +76,13 @@ func (s *Server) genDNSFilterMessage(
 // blocking mode.
 func (s *Server) genForBlockingMode(req *dns.Msg, ips []netip.Addr) (resp *dns.Msg) {
 	qt := req.Question[0].Qtype
-	switch m := s.conf.BlockingMode; m {
-	case BlockingModeCustomIP:
+	switch m := s.dnsFilter.BlockingMode; m {
+	case filtering.BlockingModeCustomIP:
 		switch qt {
 		case dns.TypeA:
-			return s.genARecord(req, s.conf.BlockingIPv4)
+			return s.genARecord(req, s.dnsFilter.BlockingIPv4)
 		case dns.TypeAAAA:
-			return s.genAAAARecord(req, s.conf.BlockingIPv6)
+			return s.genAAAARecord(req, s.dnsFilter.BlockingIPv6)
 		default:
 			// Generally shouldn't happen, since the types are checked in
 			// genDNSFilterMessage.
@@ -90,20 +90,20 @@ func (s *Server) genForBlockingMode(req *dns.Msg, ips []netip.Addr) (resp *dns.M
 
 			return s.makeResponse(req)
 		}
-	case BlockingModeDefault:
+	case filtering.BlockingModeDefault:
 		if len(ips) > 0 {
 			return s.genResponseWithIPs(req, ips)
 		}
 
 		return s.makeResponseNullIP(req)
-	case BlockingModeNullIP:
+	case filtering.BlockingModeNullIP:
 		return s.makeResponseNullIP(req)
-	case BlockingModeNXDOMAIN:
+	case filtering.BlockingModeNXDOMAIN:
 		return s.genNXDomain(req)
-	case BlockingModeREFUSED:
+	case filtering.BlockingModeREFUSED:
 		return s.makeResponseREFUSED(req)
 	default:
-		log.Error("dns: invalid blocking mode %q", s.conf.BlockingMode)
+		log.Error("dns: invalid blocking mode %q", s.dnsFilter.BlockingMode)
 
 		return s.makeResponse(req)
 	}
@@ -132,7 +132,7 @@ func (s *Server) hdr(req *dns.Msg, rrType rules.RRType) (h dns.RR_Header) {
 	return dns.RR_Header{
 		Name:   req.Question[0].Name,
 		Rrtype: rrType,
-		Ttl:    s.conf.BlockedResponseTTL,
+		Ttl:    s.dnsFilter.BlockedResponseTTL,
 		Class:  dns.ClassINET,
 	}
 }
@@ -243,6 +243,12 @@ func (s *Server) makeResponseNullIP(req *dns.Msg) (resp *dns.Msg) {
 }
 
 func (s *Server) genBlockedHost(request *dns.Msg, newAddr string, d *proxy.DNSContext) *dns.Msg {
+	if newAddr == "" {
+		log.Printf("block host is not specified.")
+
+		return s.genServerFailure(request)
+	}
+
 	ip, err := netip.ParseAddr(newAddr)
 	if err == nil {
 		return s.genResponseWithIPs(request, []netip.Addr{ip})
@@ -346,13 +352,13 @@ func (s *Server) genSOA(request *dns.Msg) []dns.RR {
 		Hdr: dns.RR_Header{
 			Name:   zone,
 			Rrtype: dns.TypeSOA,
-			Ttl:    s.conf.BlockedResponseTTL,
+			Ttl:    s.dnsFilter.BlockedResponseTTL,
 			Class:  dns.ClassINET,
 		},
 		Mbox: "hostmaster.", // zone will be appended later if it's not empty or "."
 	}
 	if soa.Hdr.Ttl == 0 {
-		soa.Hdr.Ttl = defaultValues.BlockedResponseTTL
+		soa.Hdr.Ttl = defaultBlockedResponseTTL
 	}
 	if len(zone) > 0 && zone[0] != '.' {
 		soa.Mbox += zone